Exam seater/requirements.txt
et-xmlfile==1.1.0
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
reportlab==4.0.7
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
